do not use rigid bodies in derivatives calculations. use double and not floats in derivatives calculations
git-svn-id: https://svn.salilab.org/imp/trunk@9495 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/em/include/CoarseCC.h
@@ -75,8 +75,8 @@ public:
      const DensityMap *em_map, const DensityMap *model_map,
      const Particles &model_ps,const FloatKey &mass_key,
      KernelParameters *kernel_params,
-     const float &scalefac, std::vector<float> &dvx,
-     std::vector<float>&dvy,std::vector<float>&dvz);
+     const float &scalefac, std::vector<double> &dvx,
+     std::vector<double>&dvy,std::vector<double>&dvz);
 
 
   //!Calculates the cross correlation coefficient between two maps
